HS-6 add sticky-header as module
