+++ public_html/index.html
@@ -81,7 +81,7 @@
               // Tracing
               tracesSampleRate: 1.0, // Capture 100% of the transactions
               // Session Replay
-              replaysSessionSampleRate: 1.0, // This sets the sample rate at 10%. You may want to change it to 100% while in development and then sample at a lower rate in production.
+              replaysSessionSampleRate: 0.1, // This sets the sample rate at 10%. You may want to change it to 100% while in development and then sample at a lower rate in production.
               replaysOnErrorSampleRate: 1.0, // If you're not already sampling the entire session, change the sample rate to 100% when sampling sessions where errors occur.
             });
         });
@@ -113,7 +113,7 @@
     <script src="https://www.google.com/recaptcha/api.js?render=6LeTFCAqAAAAAKlvDJZjZnVCdtD76hc3YZiIUs_Q"></script>
 
     <!-- Custom JavaScript files -->
-    <script src="/js/script.js"></script>
+    <script type="module" src="/js/script.js"></script>
     <script src="/js/sharing-buttons.js"></script>
 
     <!-- Google Tag Manager -->
