+++ src/css/style.css
@@ -241,7 +241,7 @@ nav.menu ul li.active a span {
 }
 
 /* Style dla sticky menu i nagłówka */
-header.sticky {
+header.sticky, nav.menu.sticky {
     position: fixed;
     width: 100%;
     z-index: 1000;
